+++ packages/wave/src/commands/dev/index.ts
@@ -1,0 +1,12 @@
+import { Client } from "../../util/client.js";
+
+export default async function dev(client: Client) {
+  const { output } = client;
+
+  try {
+    output.log("dev command...");
+    return 0;
+  } catch (err: unknown) {
+    output.error(err as string);
+  }
+}