+++ packages/wave/tsup.config.ts
@@ -1,0 +1,18 @@
+import { defineConfig } from "tsup";
+
+const env = process.env.NODE_ENV;
+
+export default defineConfig({
+  splitting: true,
+  clean: true, // clean up the dist folder
+  dts: true, // generate dts files
+  format: ["cjs", "esm"], // generate cjs and esm files
+  minify: env === "production",
+  bundle: env === "production",
+  skipNodeModulesBundle: true,
+  entryPoints: ["src/index.ts"],
+  watch: env === "development",
+  target: "esnext",
+  outDir: "dist",
+  entry: ["src/**/*.ts", "!src/templates/**/*"], //include all files under src
+});